Added Sample Output Files as of April 13, 2019
output_instructions.txt is the gcode file itself - feed this into the balloon assembly controller to produce expected output. This is absolutely necessary.
output_coordinates.csv is the csv file of the [x,y,z] coordinates of the end effector assembly.
output_instructions.csv is the csv file of the [X,Y,Z,E] coordinates of the tether lengths indicated in the gcode file. These last two files are really mainly for testing, to enable the other members of Project JENGA to ascertain whether or not the software aspect of the system is working as it should before the assembly of the hardware is completed (ironically, this is done by running the .csv files through software).