Update path_to_wall.py
Wrote functions for Class Basic() that make everything a bit less unclear. They'll move bricks around and adjust pickup and putdown points for new objects within a row or new rows. This is more in line with principles of OOP and breaks the code into blocks that are easier to read.
There is still some work left to do on them.